pyproject.toml
[project]
name = "doccraft"
version = "0.1.0"
description = "DocCraft: generate competition report drafts from project paper + template"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.1",
  "langgraph>=0.2.34",
  "langchain-core>=0.3.0",
  "langchain-community>=0.3.0",
  "langchain-openai>=0.2.0",
  "langchain-text-splitters>=0.3.0",
  "pypdf>=4.3.1",
  "docx2txt>=0.8",
  "python-docx>=1.1.2",
  "unstructured>=0.16.0"
]

[tool.uv]
package = false
